APIサーバーを構築してモジュール一覧の取得はAPI経由に変更
