+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+  "watch": ["src/server"],
+  "exec": "ts-node --project tsconfig.server.json src/cli/index.ts",
+  "ext": "js ts"
+}
