+++ src/cli/index.ts
@@ -1,3 +1,14 @@
-import { startServer } from "../server";
+import { createServer } from "../server";
 
-startServer();
+const port = parseInt(process.env.PORT || "3000", 10);
+const dev = process.env.NODE_ENV !== "production";
+
+createServer({ dev }).then((server) => {
+  server.listen(port, () => {
+    console.log(
+      `> Server listening at http://localhost:${port} as ${
+        dev ? "development" : "production"
+      }`
+    );
+  });
+});
