+++ src/server/index.ts
@@ -1,23 +1,31 @@
-import { createServer } from "http";
-import { parse } from "url";
 import next from "next";
+import express from "express";
+import modules from "./api/modules";
+import type { Express } from "express";
 
-const port = parseInt(process.env.PORT || "3000", 10);
-const dev = process.env.isDevelopment;
-const app = next({ dev });
-const handle = app.getRequestHandler();
+export const createServer = async ({
+  dev,
+}: {
+  dev: boolean;
+}): Promise<Express> => {
+  console.log(`dev: ${dev}`);
+  const app = next({ dev });
+  const handle = app.getRequestHandler();
 
-export const startServer = () => {
-  app.prepare().then(() => {
-    createServer((req, res) => {
-      const parsedUrl = parse(req.url!, true);
-      handle(req, res, parsedUrl);
-    }).listen(port);
+  const server = express();
 
-    console.log(
-      `> Server listening at http://localhost:${port} as ${
-        dev ? "development" : "production"
-      }`
-    );
+  try {
+    await app.prepare();
+  } catch (error) {
+    console.error(error);
+    process.exit(1);
+  }
+
+  server.use("/api/modules", modules);
+
+  server.all("*", (req, res) => {
+    return handle(req, res);
   });
+
+  return server;
 };
