+++ src/types/node.d.ts
@@ -1,6 +1,6 @@
 declare namespace NodeJS {
   interface ProcessEnv {
     NODE_ENV: "development" | "production";
-    isDevelopment: boolean;
+    isDevelopment?: boolean;
   }
 }
