+++ tsconfig.server.json
@@ -7,5 +7,5 @@
     "isolatedModules": false,
     "noEmit": false
   },
-  "include": ["src/server/**/*.ts", "src/cli/**/*.ts", "src/types/**/*.ts"]
+  "include": ["src/types/**/*.ts", "src/cli/**/*.ts", "src/server/**/*.ts"]
 }
